Refactor duplicated tests to common superclass
