+++ community/security/src/test/java/org/neo4j/server/security/auth/BasicAuthManagerTest.java
@@ -23,22 +23,18 @@ import junit.framework.TestCase;
 import org.hamcrest.core.IsEqual;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 
-import java.io.IOException;
-
+import org.neo4j.kernel.api.exceptions.InvalidArgumentsException;
+import org.neo4j.kernel.api.security.AuthManager;
 import org.neo4j.kernel.api.security.AuthSubject;
 import org.neo4j.kernel.api.security.AuthenticationResult;
-import org.neo4j.kernel.api.exceptions.InvalidArgumentsException;
 import org.neo4j.kernel.api.security.exception.InvalidAuthTokenException;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.logging.NullLogProvider;
-import org.neo4j.test.rule.fs.EphemeralFileSystemRule;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
@@ -46,16 +42,14 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.neo4j.kernel.api.security.AuthenticationResult.*;
+import static org.neo4j.kernel.api.security.AuthenticationResult.FAILURE;
+import static org.neo4j.kernel.api.security.AuthenticationResult.PASSWORD_CHANGE_REQUIRED;
+import static org.neo4j.kernel.api.security.AuthenticationResult.SUCCESS;
+import static org.neo4j.kernel.api.security.AuthenticationResult.TOO_MANY_ATTEMPTS;
 import static org.neo4j.server.security.auth.SecurityTestUtils.authToken;
 
-public class BasicAuthManagerTest
+public class BasicAuthManagerTest extends InitialUserTests
 {
-    @Rule
-    public EphemeralFileSystemRule fsRule = new EphemeralFileSystemRule();
-
-    private Config config;
-    private UserRepository users;
     private BasicAuthManager manager;
     private AuthenticationStrategy authStrategy = mock( AuthenticationStrategy.class );
 
@@ -77,105 +71,13 @@ public class BasicAuthManagerTest
     }
 
     @Test
-    public void shouldCreateDefaultUserIfNoneExist() throws Throwable
-    {
-        // When
-        manager.start();
-
-        // Then
-        final User user = users.getUserByName( "neo4j" );
-        assertNotNull( user );
-        assertTrue( user.credentials().matchesPassword( "neo4j" ) );
-        assertTrue( user.passwordChangeRequired() );
-    }
-
-    @Test
-    public void shouldLoadInitialUserIfNoneExist() throws Throwable
-    {
-        // Given
-        FileUserRepository initialUserRepository =
-                CommunitySecurityModule.getInitialUserRepository( config, NullLogProvider.getInstance(), fsRule.get() );
-        initialUserRepository.start();
-        initialUserRepository.create(
-                new User.Builder( "neo4j", Credential.forPassword( "123" ))
-                        .withRequiredPasswordChange( false )
-                        .build()
-        );
-        initialUserRepository.shutdown();
-
-        // When
-        manager.start();
-
-        // Then
-        final User user = users.getUserByName( "neo4j" );
-        assertNotNull( user );
-        assertTrue( user.credentials().matchesPassword( "123" ) );
-        assertFalse( user.passwordChangeRequired() );
-    }
-
-    @Test
-    public void shouldNotAddInitialUserIfUsersExist() throws Throwable
-    {
-        // Given
-        FileUserRepository initialUserRepository =
-                CommunitySecurityModule.getInitialUserRepository( config, NullLogProvider.getInstance(), fsRule.get() );
-        initialUserRepository.start();
-        initialUserRepository.create(
-                new User.Builder( "initUser", Credential.forPassword( "123" ))
-                        .withRequiredPasswordChange( false )
-                        .build()
-        );
-        initialUserRepository.shutdown();
-        users.start();
-        createUser( "oldUser", "321", false );
-        users.shutdown();
-
-        // When
-        manager.start();
-
-        // Then
-        final User initUser = users.getUserByName( "initUser" );
-        assertNull( initUser );
-
-        final User oldUser = users.getUserByName( "oldUser" );
-        assertNotNull( oldUser );
-        assertTrue( oldUser.credentials().matchesPassword( "321" ) );
-        assertFalse( oldUser.passwordChangeRequired() );
-    }
-
-    @Test
-    public void shouldNotUpdateUserIfInitialUserExist() throws Throwable
-    {
-        // Given
-        FileUserRepository initialUserRepository =
-                CommunitySecurityModule.getInitialUserRepository( config, NullLogProvider.getInstance(), fsRule.get() );
-        initialUserRepository.start();
-        initialUserRepository.create(
-                new User.Builder( "oldUser", Credential.forPassword( "newPassword" ))
-                        .withRequiredPasswordChange( false )
-                        .build()
-        );
-        initialUserRepository.shutdown();
-        users.start();
-        createUser( "oldUser", "oldPassword", true );
-        users.shutdown();
-
-        // When
-        manager.start();
-
-        // Then
-        final User oldUser = users.getUserByName( "oldUser" );
-        assertNotNull( oldUser );
-        assertFalse( oldUser.credentials().matchesPassword( "newPassword" ) );
-        assertTrue( oldUser.passwordChangeRequired() );
-    }
-
-    @Test
     public void shouldFindAndAuthenticateUserSuccessfully() throws Throwable
     {
         // Given
         manager.start();
-        final User user = createUser( "jake", "abc123", false );
+        User user1 = newUser( "jake", "abc123", false );
+        users.create( user1 );
+        final User user = user1;
 
         // When
         when( authStrategy.authenticate( user, "abc123" )).thenReturn( SUCCESS );
@@ -189,7 +91,9 @@ public class BasicAuthManagerTest
     {
         // Given
         manager.start();
-        final User user = createUser( "jake", "abc123", true );
+        User user1 = newUser( "jake", "abc123", true );
+        users.create( user1 );
+        final User user = user1;
 
         // When
         when( authStrategy.authenticate( user, "abc123" )).thenReturn( TOO_MANY_ATTEMPTS );
@@ -203,7 +107,9 @@ public class BasicAuthManagerTest
     {
         // Given
         manager.start();
-        final User user = createUser( "jake", "abc123", true );
+        User user1 = newUser( "jake", "abc123", true );
+        users.create( user1 );
+        final User user = user1;
 
         // When
         when( authStrategy.authenticate( user, "abc123" )).thenReturn( SUCCESS );
@@ -217,7 +123,8 @@ public class BasicAuthManagerTest
     {
         // Given
         manager.start();
-        createUser( "jake", "abc123", true );
+        User user = newUser( "jake", "abc123", true );
+        users.create( user );
 
         // Then
         assertLoginGivesResult( "unknown", "abc123", FAILURE );
@@ -313,19 +220,16 @@ public class BasicAuthManagerTest
         }
     }
 
-    private User createUser( String username, String password, boolean pwd_change )
-            throws IOException, InvalidArgumentsException
-    {
-        User user = new User.Builder( username, Credential.forPassword( password ))
-            .withRequiredPasswordChange( pwd_change ).build();
-        users.create(user);
-        return user;
-    }
-
     private void assertLoginGivesResult( String username, String password, AuthenticationResult expectedResult )
             throws InvalidAuthTokenException
     {
         AuthSubject authSubject = manager.login( authToken( username, password ) );
         assertThat( authSubject.getAuthenticationResult(), equalTo( expectedResult ) );
+    }
+
+    @Override
+    protected AuthManager authManager()
+    {
+        return manager;
     }
 }
