+++ enterprise/security/src/test/java/org/neo4j/server/security/enterprise/auth/MultiRealmAuthManagerTest.java
@@ -22,30 +22,27 @@ package org.neo4j.server.security.enterprise.auth;
 import org.apache.shiro.cache.MemoryConstrainedCacheManager;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 
 import java.util.Collections;
 
 import org.neo4j.kernel.api.exceptions.InvalidArgumentsException;
+import org.neo4j.kernel.api.security.AuthManager;
 import org.neo4j.kernel.api.security.AuthSubject;
 import org.neo4j.kernel.api.security.AuthToken;
 import org.neo4j.kernel.api.security.AuthenticationResult;
 import org.neo4j.kernel.api.security.exception.InvalidAuthTokenException;
 import org.neo4j.kernel.configuration.Config;
-import org.neo4j.server.security.enterprise.log.SecurityLog;
 import org.neo4j.kernel.impl.util.JobScheduler;
 import org.neo4j.logging.AssertableLogProvider;
 import org.neo4j.logging.Log;
 import org.neo4j.logging.NullLogProvider;
 import org.neo4j.server.security.auth.AuthenticationStrategy;
 import org.neo4j.server.security.auth.CommunitySecurityModule;
-import org.neo4j.server.security.auth.Credential;
-import org.neo4j.server.security.auth.FileUserRepository;
+import org.neo4j.server.security.auth.InitialUserTests;
 import org.neo4j.server.security.auth.PasswordPolicy;
 import org.neo4j.server.security.auth.User;
-import org.neo4j.server.security.auth.UserRepository;
-import org.neo4j.test.rule.fs.EphemeralFileSystemRule;
+import org.neo4j.server.security.enterprise.log.SecurityLog;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
@@ -63,13 +60,8 @@ import static org.neo4j.logging.AssertableLogProvider.inLog;
 import static org.neo4j.server.security.auth.SecurityTestUtils.authToken;
 import static org.neo4j.test.assertion.Assert.assertException;
 
-public class MultiRealmAuthManagerTest
+public class MultiRealmAuthManagerTest extends InitialUserTests
 {
-    @Rule
-    public EphemeralFileSystemRule fsRule = new EphemeralFileSystemRule();
-
-    private Config config;
-    private UserRepository users;
     private AuthenticationStrategy authStrategy;
     private MultiRealmAuthManager manager;
     private EnterpriseUserManager userManager;
@@ -114,92 +106,6 @@ public class MultiRealmAuthManagerTest
     {
         manager.stop();
         manager.shutdown();
-    }
-
-    @Test
-    public void shouldCreateDefaultUserIfNoneExist() throws Throwable
-    {
-        // When
-        manager.start();
-
-        // Then
-        final User user = users.getUserByName( "neo4j" );
-        assertNotNull( user );
-        assertTrue( user.credentials().matchesPassword( "neo4j" ) );
-        assertTrue( user.passwordChangeRequired() );
-    }
-
-    @Test
-    public void shouldLoadInitialUserIfNoneExist() throws Throwable
-    {
-        // Given
-        FileUserRepository initialUserRepository =
-                CommunitySecurityModule.getInitialUserRepository( config, NullLogProvider.getInstance(), fsRule.get() );
-        initialUserRepository.start();
-        initialUserRepository.create(
-                new User.Builder( "neo4j", Credential.forPassword( "123" ))
-                        .withRequiredPasswordChange( false )
-                        .build()
-        );
-        initialUserRepository.shutdown();
-
-        // When
-        manager.start();
-
-        // Then
-        final User user = users.getUserByName( "neo4j" );
-        assertNotNull( user );
-        assertTrue( user.credentials().matchesPassword( "123" ) );
-        assertFalse( user.passwordChangeRequired() );
-    }
-
-    @Test
-    public void shouldNotAddInitialUserIfUsersExist() throws Throwable
-    {
-        // Given
-        FileUserRepository initialUserRepository =
-                CommunitySecurityModule.getInitialUserRepository( config, NullLogProvider.getInstance(), fsRule.get() );
-        initialUserRepository.start();
-        initialUserRepository.create( newUser( "initUser", "123", false ) );
-        initialUserRepository.shutdown();
-        users.start();
-        users.create( newUser( "oldUser", "321", false ) );
-        users.shutdown();
-
-        // When
-        manager.start();
-
-        // Then
-        final User initUser = users.getUserByName( "initUser" );
-        assertNull( initUser );
-
-        final User oldUser = users.getUserByName( "oldUser" );
-        assertNotNull( oldUser );
-        assertTrue( oldUser.credentials().matchesPassword( "321" ) );
-        assertFalse( oldUser.passwordChangeRequired() );
-    }
-
-    @Test
-    public void shouldNotUpdateUserIfInitialUserExist() throws Throwable
-    {
-        // Given
-        FileUserRepository initialUserRepository =
-                CommunitySecurityModule.getInitialUserRepository( config, NullLogProvider.getInstance(), fsRule.get() );
-        initialUserRepository.start();
-        initialUserRepository.create( newUser( "oldUser", "newPassword", false ) );
-        initialUserRepository.shutdown();
-        users.start();
-        users.create( newUser( "oldUser", "oldPassword", true ) );
-        users.shutdown();
-
-        // When
-        manager.start();
-
-        // Then
-        final User oldUser = users.getUserByName( "oldUser" );
-        assertNotNull( oldUser );
-        assertTrue( oldUser.credentials().matchesPassword( "oldPassword" ) );
-        assertTrue( oldUser.passwordChangeRequired() );
     }
 
     @Test
@@ -714,16 +620,15 @@ public class MultiRealmAuthManagerTest
         return inLog( this.getClass() ).error( message, (Object[]) arguments );
     }
 
-    private User newUser( String userName, String password, boolean pwdChange )
-    {
-        return new User.Builder( userName, Credential.forPassword( password ))
-                    .withRequiredPasswordChange( pwdChange )
-                    .build();
-    }
-
     private void setMockAuthenticationStrategyResult( String username, String password, AuthenticationResult result )
     {
         final User user = users.getUserByName( username );
         when( authStrategy.authenticate( user, password ) ).thenReturn( result );
+    }
+
+    @Override
+    protected AuthManager authManager()
+    {
+        return manager;
     }
 }
